[chore] implement feedback
