+++ .eslintrc
@@ -23,6 +23,7 @@
     "new-cap": 0,
     "consistent-return": 0,
     "arrow-parens": ["error", "as-needed"],
+    "react/display-name": 1,
     "no-param-reassign": 0,
     "comma-dangle": 0,
     "curly": ["error", "multi-line"],
