+++ src/components/presentationals/AuthenticationCard/AuthenticationCard.scss
@@ -1,4 +1,4 @@
-@import "../../assets/variables";
+@import "../../../assets/variables";
 
 .card {
   background: $primary-white;
