+++ src/components/presentationals/AuthenticationIllustration/AuthenticationIllustration.scss
