+++ src/components/presentationals/Avatar/Avatar.scss
