+++ src/components/presentationals/Button/Button.scss
@@ -1,5 +1,5 @@
 @import url("https://fonts.googleapis.com/css?family=Lato:400,700|Lora:400,700|Sacramento");
-@import "../../assets/variables";
+@import "../../../assets/variables";
 
 .btn {
   background-color: $secondary-grey;
@@ -10,8 +10,12 @@
   font-size: 14px;
   height: 50px;
   letter-spacing: 0.1em;
+<<<<<<< HEAD:src/components/presentationals/Button.scss
   text-align: center;
   width: 100%;
+=======
+  cursor: pointer;
+>>>>>>> [chore] implement feedback:src/components/presentationals/Button/Button.scss
 }
 
 .btn-dark {
