+++ src/components/presentationals/Navbar/Navbar.jsx
@@ -1,9 +1,9 @@
+/* eslint-disable no-console */
 import React from 'react';
 import PropTypes from 'prop-types';
 
 import { Dropdown } from 'semantic-ui-react';
-import Avatar from './Avatar.jsx';
-import './Avatar.jsx';
+import Avatar from '../Avatar/Avatar';
 import './Navbar.scss';
 
 const Navbar = ({ props }) => {
@@ -13,7 +13,7 @@ const Navbar = ({ props }) => {
         <div>
           <h1 className="brand">Authors Haven</h1>
         </div>
-        {props.profileUrl !== '' && (
+        {props.profileUrl && (
           <div className="navbar-avatar">
             <Avatar avatarUrl={props.profileUrl} />
             <Dropdown className="dropdown">
