+++ src/components/presentationals/ResetPassword/ResetPassword.jsx
@@ -1,24 +1,23 @@
 import React from 'react';
 import { Form } from 'semantic-ui-react';
 
-import Title from './Title.jsx';
-import Button from './Button.jsx';
-import './AuthenticationCard.scss';
-import './Form.scss';
+import Title from '../Title/Title';
+import Button from '../Button/Button';
+import '../AuthenticationCard/AuthenticationCard.scss';
+import '../Form.scss';
 
 const ForgotPasswordCard = () => {
   return (
     <div className="right">
-      <div className="card forgot-password">
+      <div className="card reset-password">
         <Title title={title} />
         <hr />
-        <p>
-          Enter the email you used to sign up and we‘ll send you a password
-          reset link
-        </p>
         <Form>
           <Form.Field>
-            <input placeholder="Email" type="email" />
+            <input placeholder="New Password" type="password" />
+          </Form.Field>
+          <Form.Field>
+            <input placeholder="Confirm Password" type="password" />
           </Form.Field>
           <Button button={buttonValues} />
         </Form>
@@ -28,13 +27,13 @@ const ForgotPasswordCard = () => {
 };
 
 const title = {
-  content: 'Forgot Password?',
+  content: 'Reset Password',
   class: 'text-center',
 };
 
 const buttonValues = {
   type: 'submit',
-  value: 'SEND EMAIL',
+  value: 'RESET PASSWORD',
   class: 'btn-dark',
 };
