--- src/store/actions/types.js
@@ -1,3 +1,0 @@
-export const SIGNUP_SUCCESS = 'SIGNUP_SUCCESS';
-export const SIGNUP_ERROR = 'SIGNUP_ERROR';
-export const SIGNUP_REQUESTED = 'SIGNUP_REQUESTED';
