+++ src/store/rootReducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
 
-import auth from './authReducer';
+import { auth } from './modules/auth';
 
 export default combineReducers({ auth });
