+++ src/utils/formValidator.js
@@ -1,5 +1,3 @@
-/* eslint-disable no-unused-vars */
-/* eslint-disable no-console */
 export default function validator(userCredentials) {
   const validationErrors = [];
   Object.keys(userCredentials).forEach(key => {
